fast_planner/plan_manage : debugging cpp code
We were facing the error roscpp invalid argument, the root cause is yet
not known so we are procedding now step by step to run and test each node
individually and comment the other just to debug.

In this commit we commented fast planner and just running the entrypoint of
fastplanner  and from there entry point is kinofsm node and from there
entry point is traj_utils. Apart from entry point functions all other code is
commented out

Signed-off-by: charvi-077 <charvi077@gmail.com>